Initial import of a client for Octopus Energy APIs.
Note that while this currently only implements REST APIs, it is meant to
support the GraphQL APIs as well, as some of the information is only
available for it.

Also, this currently only implements electricity consumption because I
don't have an easy way to make sure the gas consumption works as
documented.